requirements.txt
alembic==1.12.0
aniso8601==9.0.1
anytree==2.8.0
attrs==23.1.0
backports.tempfile==1.0
backports.weakref==1.0.post1
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.7
deprecation==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
flask-restx==1.2.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.26.0
oauthlib==3.2.2
packaging==23.1
pandas==1.3.3
pikepdf==8.4.1
Pillow==10.0.1
preppy==4.0.0
psycopg2-binary==2.9.8
pydot==1.4.2
Pygments==2.16.1
PyJWT==1.7.1
pyparsing==3.1.1
pyrsistent==0.20.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
reportlab==3.6.2
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
SQLAlchemy==1.4.27
sqlalchemy-schemadisplay==1.3
typing_extensions==4.8.0
urllib3==1.26.16
Werkzeug==2.0.3
z3c.rml==4.2.0
zope.event==5.0
zope.interface==6.0
zope.schema==7.0.1
